feat: admit by lua

+++ src/main/kotlin/taskforce/fcfs/allocate/QueueAdmitProperties.kt
@@ -1,5 +1,5 @@
 package taskforce.fcfs.allocate
 
 interface QueueAdmitProperties {
-    fun getRequest(): Int
+    fun getRequest(): Long
 }
