+++ src/main/kotlin/taskforce/fcfs/clientqueue/YamlEventProperties.kt
@@ -1,6 +1,5 @@
 package taskforce.fcfs.clientqueue
 
-import taskforce.fcfs.clientqueue.EventProperties
 import org.springframework.stereotype.Component
 
 // TODO config 와 return bean 으로 관리해서 유연성 확보 필요
@@ -8,5 +7,5 @@ import org.springframework.stereotype.Component
 class YamlEventProperties : EventProperties {
 
     override fun getEventName(): String  = "CHICKEN"
-    override fun getEventLimit(): Int = 1000
+    override fun getEventLimit(): Long = 1000L
 }
